Componentize the product form location field
